refactor: Notification에서 NotificationHandler로 클래스명 수정 및 메서드명 개선
