+++ app/src/main/java/woowacourse/movie/notification/NotificationReceiver.kt
@@ -10,10 +10,10 @@ import android.content.Intent
 import androidx.core.app.NotificationCompat
 import woowacourse.movie.MovieApplication.Companion.sharedPreferences
 import woowacourse.movie.R
-import woowacourse.movie.notification.Notification.Companion.NOTIFICATION_ID
-import woowacourse.movie.notification.Notification.Companion.PENDING_REQUEST_CODE
-import woowacourse.movie.notification.Notification.Companion.PUT_EXTRA_KEY_MOVIE_TITLE_ID
-import woowacourse.movie.notification.Notification.Companion.PUT_EXTRA_KEY_RESERVATION_ID
+import woowacourse.movie.notification.NotificationHandler.Companion.NOTIFICATION_ID
+import woowacourse.movie.notification.NotificationHandler.Companion.PENDING_REQUEST_CODE
+import woowacourse.movie.notification.NotificationHandler.Companion.PUT_EXTRA_KEY_MOVIE_TITLE_ID
+import woowacourse.movie.notification.NotificationHandler.Companion.PUT_EXTRA_KEY_RESERVATION_ID
 import woowacourse.movie.presentation.ui.main.MainActivity
 import woowacourse.movie.presentation.ui.reservation.ReservationActivity
